Refactor: 컴포넌트 별 모듈화
