+++ src/js/components/WinningNumberForm.js
@@ -1,0 +1,3 @@
+export default class WinningNumberForm {
+  constructor() {}
+}
